change how we include auxiliary classes. (Stacktrace, Globals, and MD5 are now ExtJS Classes and will be included in the build app.js) This also fixes the Selenium error with mathjax.
git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtWebApp/trunk@4309 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ test-ui.sh
@@ -9,19 +9,36 @@ export SELENIUM_DRIVER=*googlechrome
 
 export SELENIUM_JAR=~/Applications/bin/selenium-server-standalone-2.16.1.jar
 
-#port for the simpleserver
 export PORT=8181
 
-# fire up an http server in the background
-echo "Starting SimpleHTTP Server"
-python -m SimpleHTTPServer $PORT >/dev/null 2>&1 &
+if [[ ! -z $FLAGGED ]] ; then
+	export TEST_URL=http://localhost/index.html
+	#do dataserver setup...
 
+else #just launch assuming everything is running
+	echo "Starting SimpleHTTP Server"
+
+	python -m SimpleHTTPServer $PORT >/dev/null 2>&1 &
+
+	HPID=`jobs -l 1 | awk '{print $2}'`
+	export TEST_URL=http://localhost:$PORT/index.html
+fi
+
+
+echo "Running tests at $TEST_URL"
+
+#run tests
 nosetests-2.7 -v -d -w ./src/test/python/
 
-# kill the http server
-echo "Stopping Simple HTTP Server"
-HPID=`jobs -l 1 | awk '{print $2}'`
-kill -9 $HPID
+
+
+#shutdown/cleanup
+
+if [[ ! -z $HPID ]] ; then
+	# kill the http server
+	echo "Stopping SimpleHTTP Server"
+	kill -9 $HPID
+fi
